Includes memberships without an event to correctly count participants
The join table created by the left join between field_test_memberships
and field_test_events will not contain an event name for participants
who did not convert and therefore do not have an associated event.

Including `nil` in the portion of the query used to scope to a given
event name ensures that we will also include participants who did not
convert.